[FIX] l10n_ar: fix table size in pdf form
The tax breakdown is required by Argentinian law to be displayed in the
bottom left corner of the invoice report. This was added in [^1] and
works well for the HTML view but min-width related classes don't work
well with pdf rendering, so the table stretched across the entire width
of the pdf.

The fix is to follow the rest of the file and use a fixed col width on
the table for the pdf view.

[^1]: https://github.com/odoo/odoo/pull/201257
task-4779975

closes odoo/odoo#214696

X-original-commit: 1a21f4eb71111b2850b18d7efc53ae6d3e3410da
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Andrew Gavgavian (andg) <andg@odoo.com>